+++ src/FFmpeg/ffcoloritem.cpp
@@ -2,37 +2,73 @@
 
 FFColorItem::FFColorItem(QString name, QObject *parent) : FFBaseObject(name, parent)
 {
-    _scaleName="";
+    _outputScaleName="";
     _scaleFilter = ZScale;
+    _abilities.setFlag(Input,true);
+    _abilities.setFlag(Output,true);
+    _metadataName = name;
+    _inputScaleName = "";
 }
 
 FFColorItem::FFColorItem(QString name, QString prettyName, QObject *parent) : FFBaseObject(name, prettyName, parent)
 {
-    _scaleName="";
+    _outputScaleName="";
     _scaleFilter = ZScale;
+    _abilities.setFlag(Input,true);
+    _abilities.setFlag(Output,true);
+    _metadataName = name;
+    _inputScaleName = "";
 }
 
-FFColorItem::FFColorItem(QString name, QString prettyName, QString scaleName, QObject *parent) : FFBaseObject(name, prettyName, parent)
+FFColorItem::FFColorItem(QString name, QString prettyName, QString outputScaleName, QObject *parent) : FFBaseObject(name, prettyName, parent)
 {
-    _scaleName = scaleName;
+    _outputScaleName = outputScaleName;
     _scaleFilter = ZScale;
+    _abilities.setFlag(Input,true);
+    _abilities.setFlag(Output,true);
+    _metadataName = name;
+    _inputScaleName = "";
 }
 
-FFColorItem::FFColorItem(QString name, QString prettyName, QString scaleName, ScaleFilter scaleFilter, QObject *parent) : FFBaseObject(name, prettyName, parent)
+FFColorItem::FFColorItem(QString name, QString prettyName, QString outputScaleName, ScaleFilter scaleFilter, QObject *parent) : FFBaseObject(name, prettyName, parent)
 {
-    _scaleName = scaleName;
+    _outputScaleName = outputScaleName;
     _scaleFilter = scaleFilter;
+    _abilities.setFlag(Input,true);
+    _abilities.setFlag(Output,true);
+    _metadataName = name;
+    _inputScaleName = "";
+}
+
+bool FFColorItem::isInput() const
+{
+    return _abilities.testFlag(Input);
+}
+
+bool FFColorItem::isOutput() const
+{
+    return _abilities.testFlag(Output);
+}
+
+void FFColorItem::setInput(bool i)
+{
+    _abilities.setFlag(Input,i);
+}
+
+void FFColorItem::setOutput(bool o)
+{
+    _abilities.setFlag(Output,o);
 }
 
 
-QString FFColorItem::scaleName() const
+QString FFColorItem::outputScaleName() const
 {
-    return _scaleName;
+    return _outputScaleName;
 }
 
-void FFColorItem::setScaleName(const QString &zscaleName)
+void FFColorItem::setOutputScaleName(const QString &zscaleName)
 {
-    _scaleName = zscaleName;
+    _outputScaleName = zscaleName;
 }
 
 FFColorItem::ScaleFilter FFColorItem::scaleFilter() const
@@ -43,4 +79,24 @@ FFColorItem::ScaleFilter FFColorItem::scaleFilter() const
 void FFColorItem::setScaleFilter(const ScaleFilter &scaleFilter)
 {
     _scaleFilter = scaleFilter;
+}
+
+QString FFColorItem::metadataName() const
+{
+    return _metadataName;
+}
+
+void FFColorItem::setMetadataName(const QString &metadataName)
+{
+    _metadataName = metadataName;
+}
+
+QString FFColorItem::inputScaleName() const
+{
+    return _inputScaleName;
+}
+
+void FFColorItem::setInputScaleName(const QString &inputScaleName)
+{
+    _inputScaleName = inputScaleName;
 }