+++ src/FFmpeg/fflut.h
@@ -2,6 +2,7 @@
 #define FFLUT_H
 
 #include "ffbaseobject.h"
+#include "Renderer/cachemanager.h"
 
 class FFLut : public FFBaseObject
 {
@@ -9,28 +10,22 @@ public:
     enum Type{ OneD, ThreeD };
     Q_ENUM(Type)
 
-    FFLut(QString name, QString prettyName, QString inputTrc = "", QString outputTrc = "", QObject *parent = nullptr);
+    FFLut(QString name, QString prettyName, QString inputProfile = "", QString outputProfile = "", Type type = ThreeD, QObject *parent = nullptr);
 
     Type type() const;
     void setType(const Type &type);
 
-    QString outputTrc() const;
-    void setOutputTrc(const QString &outputTrc);
+    QString outputProfile() const;
+    void setOutputProfile(const QString &outputProfile);
 
-    QString inputTrc() const;
-    void setInputTrc(const QString &inputTrc);
+    QString inputProfile() const;
+    void setInputProfile(const QString &inputProfile);
 
-    QString outputPrimaries() const;
-    void setOutputPrimaries(const QString &outputPrimaries);
-
-    QString inputPrimaries() const;
-    void setInputPrimaries(const QString &inputPrimaries);
+    QString extract();
 
 private:
-    QString _inputPrimaries;
-    QString _outputPrimaries;
-    QString _inputTrc;
-    QString _outputTrc;
+    QString _inputProfile;
+    QString _outputProfile;
     Type _type;
 };