+++ src/UI/Blocks/blockcolor.cpp
@@ -9,19 +9,19 @@ BlockColor::BlockColor(MediaInfo *mediaInfo, QWidget *parent) :
 
     //pouplate boxes
     FFmpeg *ffmpeg = FFmpeg::instance();
-    foreach(FFBaseObject *c, ffmpeg->colorPrimaries())
+    foreach(FFColorItem *c, ffmpeg->colorPrimaries())
     {
         primariesBox->addItem( c->prettyName(), c->name() );
     }
-    foreach(FFBaseObject *c, ffmpeg->colorTRCs())
+    foreach(FFColorItem *c, ffmpeg->colorTRCs())
     {
         trcBox->addItem( c->prettyName(), c->name() );
     }
-    foreach(FFBaseObject *c, ffmpeg->colorSpaces())
+    foreach(FFColorItem *c, ffmpeg->colorSpaces())
     {
         spaceBox->addItem( c->prettyName(), c->name() );
     }
-    foreach(FFBaseObject *c, ffmpeg->colorRanges())
+    foreach(FFColorItem *c, ffmpeg->colorRanges())
     {
         rangeBox->addItem( c->prettyName(), c->name() );
     }